fix: restore closeout-settings CI contract and line ratchets
- Test cells that stub settings gain readRepository (complete path reads
  repository closeout settings through it).
- Review/consent/fact-disposition flow tests pin the strict profile in
  their fixtures; the standard default keeps those gates off by design.
- GUI settings fixtures carry the now-required closeout field.
- Split the closeout settings domain into settings-closeout.ts so
  settings.ts returns under its line baseline; register the new module
  in the pinned offline-maintenance closure.
- Merge the duplicate walFlush/reviewReturnBudget and boolean payload
  checks in daemon-protocol-rpc-validation.ts back to the file baseline;
  the closeout profile enum stays kernel-judged.
- Extract the closeout settings rows from SettingsView into
  views/settings/CloseoutRows.tsx with the shared SettingSelect widget.
