+++ packages/gui/test/settings-repository-selectors.vitest.ts
@@ -22,6 +22,7 @@ const SETTINGS = {
   defaultPreset: "standard-task",
   defaultProfile: "baseline",
   reviewIndependence: "execution" as const,
+  closeout: { profile: "standard" as const },
   locale: "zh-CN" as const,
   scaffolds: { task: "governance/task-scaffold.json", repository: "governance/repository-scaffold.json" },
   walFlush: { adaptive: true, events: 256, bytes: 8_388_608, milliseconds: 3_600_000 },
