+++ packages/gui/test/system-group-widescreen.vitest.ts
@@ -46,6 +46,7 @@ function seedQueries(client: QueryClient): void {
       defaultPreset: "standard-task",
       defaultProfile: "baseline",
       reviewIndependence: "execution",
+      closeout: { profile: "standard" },
       locale: "zh-CN",
       scaffolds: { task: "governance/task-scaffold.json", repository: "governance/repository-scaffold.json" },
       walFlush: { adaptive: true, events: 256, bytes: 8_388_608, milliseconds: 3_600_000 },
@@ -281,6 +282,7 @@ describe("Settings kind renderer consumes and updates the daemon-owned facet", (
         defaultVertical: "software/coding",
         defaultPreset: "standard-task",
         defaultProfile: "baseline",
+        closeout: { profile: "standard" as const },
         locale: "zh-CN" as const,
         scaffolds: { task: "governance/task-scaffold.json", repository: "governance/repository-scaffold.json" },
         walFlush: { adaptive: true, events: 256, bytes: 8_388_608, milliseconds: 3_600_000 },
