+++ packages/kernel/src/domain/index.ts
@@ -270,9 +270,9 @@ export {
   validateRepositorySettings,
   validateSettingsV1,
   writeRepositorySettingsFacet,
-  effectiveCloseoutGates,
 } from "./settings.ts";
 export type { RepositorySettingsV1, SettingsLocale, SettingsV1 } from "./settings.ts";
+export { effectiveCloseoutGates } from "./settings-closeout.ts";
 export { compileSettingsChangedEvent, isSettingsEvent } from "./settings-event.ts";
 export {
   applyVerticalKindCommand,
