+++ packages/kernel/src/domain/settings.ts
@@ -1,6 +1,14 @@
 import { setting, settingBlockValue } from "../layout/harness-settings.ts";
 import type { EntityDocumentJsonSchema } from "./entity-json-schema.ts";
 import { validateEntityJsonSchema } from "./entity-json-schema.ts";
+import {
+  DEFAULT_CLOSEOUT_SETTINGS,
+  closeoutOverrideKeys,
+  closeoutProfiles,
+  readCloseoutSettings,
+  writeCloseoutFacet,
+  type CloseoutSettingsV1,
+} from "./settings-closeout.ts";
 
 export const SETTINGS_ID = "repository";
 export const SETTINGS_LOCAL_PATH = ".harness/settings.local.json";
@@ -10,29 +18,6 @@ export const reviewIndependenceLevels = ["execution", "principal"] as const;
 export type ReviewIndependence = (typeof reviewIndependenceLevels)[number];
 export const DEFAULT_RESTORE_DRILL_RETENTION = 3;
 export const DEFAULT_CI_WORKFLOWS = Object.freeze([] as const);
-export const closeoutProfiles = ["standard", "strict"] as const;
-export type CloseoutProfile = (typeof closeoutProfiles)[number];
-export const closeoutOverrideKeys = ["review", "consent", "factDisposition", "codeDoc"] as const;
-export type CloseoutOverrideKey = (typeof closeoutOverrideKeys)[number];
-export type CloseoutOverridesV1 = Readonly<Partial<Record<CloseoutOverrideKey, boolean>>>;
-export interface CloseoutSettingsV1 {
-  readonly profile: CloseoutProfile;
-  readonly overrides?: CloseoutOverridesV1;
-}
-export type CloseoutGate = CloseoutOverrideKey;
-
-export function effectiveCloseoutGates(
-  closeout: CloseoutSettingsV1,
-  taskGateIds: readonly string[] = [],
-): Readonly<Record<CloseoutGate, boolean>> {
-  const baseline = closeout.profile === "strict";
-  return Object.freeze({
-    review: closeout.overrides?.review ?? baseline,
-    consent: closeout.overrides?.consent ?? baseline,
-    factDisposition: closeout.overrides?.factDisposition ?? baseline,
-    codeDoc: (closeout.overrides?.codeDoc ?? baseline) || taskGateIds.includes("code-doc-reconciliation"),
-  });
-}
 
 export interface WalFlushSettingsV1 {
   readonly adaptive: boolean;
@@ -147,7 +132,7 @@ export const INITIAL_SETTINGS_V1: SettingsV1 = Object.freeze({
   }),
   walFlush: DEFAULT_WAL_FLUSH_SETTINGS,
   ci: Object.freeze({ workflows: DEFAULT_CI_WORKFLOWS }),
-  closeout: Object.freeze({ profile: "standard" }),
+  closeout: DEFAULT_CLOSEOUT_SETTINGS,
   restoreDrillRetention: DEFAULT_RESTORE_DRILL_RETENTION,
 });
 
@@ -461,23 +446,6 @@ function closeoutSettingsSchema() {
   };
 }
 
-function readCloseoutSettings(body: string): CloseoutSettingsV1 {
-  const profile = settingBlockValue(body, "closeout", "profile");
-  if (profile === undefined) return INITIAL_SETTINGS_V1.closeout;
-  if (!closeoutProfiles.includes(profile as CloseoutProfile))
-    throw new Error(`settings.closeout.profile must be one of ${closeoutProfiles.join(", ")}`);
-  const section = /^  closeout:[^\S\r\n]*(?:\r?\n)((?:(?:    |      )[^\r\n]*(?:\r?\n|$))*)/mu.exec(body)?.[1] ?? "";
-  const overrides = Object.fromEntries(
-    closeoutOverrideKeys.flatMap((key) => {
-      const raw = new RegExp(`^      ${key}:[^\\S\\r\\n]*([^#\\r\\n]*?)\\s*(?:#.*)?$`, "mu").exec(section)?.[1]?.trim();
-      if (raw === undefined) return [];
-      if (raw !== "true" && raw !== "false") throw new Error(`settings.closeout.overrides.${key} must be boolean`);
-      return [[key, raw === "true"]];
-    }),
-  ) as CloseoutOverridesV1;
-  return { profile: profile as CloseoutProfile, ...(Object.keys(overrides).length ? { overrides } : {}) };
-}
-
 function readCiSettings(body: string): SettingsV1["ci"] {
   const section = /^  ci:[^\S\r\n]*(?:\r?\n)((?:    [^\r\n]*(?:\r?\n|$))*)/mu.exec(body)?.[1];
   if (section === undefined) return INITIAL_SETTINGS_V1.ci;
@@ -550,25 +518,6 @@ function writeCiFacet(body: string, ci: RepositorySettingsV1["ci"]): string {
     isDefault = JSON.stringify(ci) === JSON.stringify(INITIAL_SETTINGS_V1.ci);
   if (!section.test(body) && isDefault) return body;
   const rendered = `  ci:\n    workflows: [${ci.workflows.join(", ")}]\n`;
-  if (section.test(body)) return body.replace(section, rendered);
-  const header = /^settings:[^\r\n]*(?:\r?\n|$)/mu;
-  if (!header.test(body)) throw new Error("Missing settings block in harness.yaml.");
-  return body.replace(header, (match) => `${match}${rendered}`);
-}
-
-function writeCloseoutFacet(body: string, closeout: CloseoutSettingsV1): string {
-  const section = /^  closeout:[^\S\r\n]*(?:\r?\n)(?:(?:    |      )[^\r\n]*(?:\r?\n|$))*/mu,
-    isDefault = JSON.stringify(closeout) === JSON.stringify(INITIAL_SETTINGS_V1.closeout);
-  if (!section.test(body) && isDefault) return body;
-  const overrideLines = closeoutOverrideKeys.flatMap((key) =>
-      closeout.overrides?.[key] === undefined ? [] : [`      ${key}: ${closeout.overrides[key]}`],
-    ),
-    rendered = [
-      "  closeout:",
-      `    profile: ${closeout.profile}`,
-      ...(overrideLines.length ? ["    overrides:", ...overrideLines] : []),
-      "",
-    ].join("\n");
   if (section.test(body)) return body.replace(section, rendered);
   const header = /^settings:[^\r\n]*(?:\r?\n|$)/mu;
   if (!header.test(body)) throw new Error("Missing settings block in harness.yaml.");
